Refs #34322 - move TracesTab
