+++ webpack/components/extensions/HostDetails/Tabs/TracesTab/TracesTab.scss
